% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits2DT.R
\name{addField2Hits}
\alias{addField2Hits}
\title{A function that adds the field values to a \code{featureHits} object.}
\usage{
addField2Hits(hits, tc)
}
\arguments{
\item{hits}{A \code{featureHits} object from the corpustools package.}

\item{tc}{A \code{tCorpus} object that was built using the same dataframe that is used for \code{efoDF}}
}
\value{
A \code{featureHits} object a column added to the \code{hits} component providing information on which
field (subject, synonym or match) the hit was found in.
}
\description{
Users query the corpus to find EFO ontology terms that match the query. This function adds the \code{field} information
from the \code{tCorpus} to the \code{featureHits} object.
}
\details{
The hits are aligned with the appropriate rows in the \code{tc$tokens} data.frame and
the \code{field} values extracted and entered added to the \code{hits} dataframe of the \code{hits} object.
}
\examples{
hits <- search_features(efo_tc, query = c("Infect# infectious*", "Pancreas# pancrea*"))
summary(hits)
hitswF <- addField2Hits(hits, efo_tc)
}
\seealso{
\code{\link[corpustools:tCorpus]{corpustools::tCorpus()}}, \link{efo_df}
}
\author{
R. Gentleman
}
