% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocgr_from_db.R
\name{assocgr_from_db}
\alias{assocgr_from_db}
\title{build a GRanges with all GWAS associations from SQLite db}
\usage{
assocgr_from_db()
}
\description{
build a GRanges with all GWAS associations from SQLite db
}
\note{
Uses gwascatalog_associations table in the GWASCatalogSearchDB.
}
