% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efo_oi}
\alias{efo_oi}
\title{a static instance of an ontology_index representing the
EFO "is-a" content from the SQLite database}
\format{
ontologyIndex ontology_index instance
}
\usage{
efo_oi
}
\description{
a static instance of an ontology_index representing the
EFO "is-a" content from the SQLite database
}
\note{
Should be regenerated and stored whenever EFO content of the
database changes.
}
\examples{
data("efo_oi", package="gwasCatSearch")
efo_oi
}
\keyword{datasets}
