% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efo_tc}
\alias{efo_tc}
\title{A tCorpus object built using EFO labels.}
\format{
a tCorpus R6 object where documents are EFO labels suitable for searching
}
\source{
The data are created from \code{efo_df}
}
\usage{
efo_tc
}
\description{
A tCorpus object built using EFO labels.
}
\examples{
efo_tc
ht2 <- search_features(efo_tc, query = "infectious* AND pancrea*")
summary(ht2)
}
\references{
EFO: https://www.ebi.ac.uk/efo/
}
\seealso{
\link{efo_df}, \code{\link[corpustools:tCorpus]{corpustools::tCorpus()}}
}
\keyword{datasets}
