% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{EFOchildren}
\alias{EFOchildren}
\title{A function to query the efo_edges table and return the child terms for the input EFO IDs}
\usage{
EFOchildren(EFOID)
}
\arguments{
\item{EFOID}{a character vector of the EFO CURIE symbols}
}
\value{
A named vector, names are from the input EFOIDs and the values are the corresponding CURIEs for the child terms.
}
\description{
This function provides an interface to the SQL database containing EFO edges (parent child relationships).
}
\details{
The function returns the set of children (more specific) for the input IDs.
}
\examples{
EFOchildren("EFO:0009448")  ##should contain EFO:0000768
EFOchildren(c("EFO:0000768", "MONDO:0002429"))
}
\author{
Robert Gentleman
}
