% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{EFOdescendants}
\alias{EFOdescendants}
\title{A function to query the efo_edges table and return the descendant terms for the input EFO IDs}
\usage{
EFOdescendants(EFOID)
}
\arguments{
\item{EFOID}{a character vector of the EFO CURIE symbols}
}
\value{
A named vector, names are from the input EFOIDs and the values are the corresponding CURIEs for the
descendant terms.  These are basically the children, the children of the children and so on.
We use entailed edges so there may be terms in the descendents that have been identified
by logic based on the ontology and may not be detected by the recursive application of EFOchildren.
}
\description{
This function provides an interface to the SQL database containing EFO entailed edges.
}
\details{
The function returns the set of descendents for the input terms.
}
\examples{
##leukemia
dec1 = EFOdescendants("EFO:0000565")  
table(names(dec1))
dec2 = EFOdescendants(c("EFO:0000768", "MONDO:0002429"))
}
\author{
Robert Gentleman
}
