% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{EFOlabels}
\alias{EFOlabels}
\title{A function to query the efo_labels table and return the term labels for the input EFO IDs}
\usage{
EFOlabels(EFOID)
}
\arguments{
\item{EFOID}{a character vector of the EFO CURIE symbols}
}
\description{
This function provides an interface to the SQL database containing EFO entailed edges.
}
\details{
The function returns the set of descendents for the input terms.
}
\examples{
EFOlabels(c("EFO:0000768", "MONDO:0002429"))
}
