% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_variant_context.R
\name{get_variant_context}
\alias{get_variant_context}
\title{utility for extracting information about GWAS hits near a position,
uses a snapshot of EBI GWAS catalog}
\usage{
get_variant_context(chr = 15, pos = 6.9e+07, radius = 5e+05, gwcat)
}
\arguments{
\item{chr}{chromosome identifier}

\item{pos}{numeric(1) position of variant of interest}

\item{radius}{numeric(1) radius of interval to present}

\item{gwcat}{instance of gwaswloc from gwascat package}
}
\value{
a data.frame with fields PVALUE_MLOG, STUDY.ACCESSION, MAPPED_TRAIT,
SNP_ID_CURRENT, PUBMEDID, pos, mlogp, and tag, the latter an HTML string
for use in tooltips
}
\description{
utility for extracting information about GWAS hits near a position,
uses a snapshot of EBI GWAS catalog
}
\examples{
data(gwascat_2023_06_24)
mydf = get_variant_context(gwcat = gwascat_2023_06_24)
head(mydf[, 1:6])
}
