% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getDiseaseLocation}
\alias{getDiseaseLocation}
\title{A function to return the disease locations from the efo_labels table.}
\usage{
getDiseaseLocation(Ontonames)
}
\arguments{
\item{Ontonames}{a character vector of the ontology CURIE symbols}
}
\value{
A named vector of disease locations.
}
\description{
This function takes an input vector of ontology CURIE symbols and returns the inferred disease locations.
}
\details{
The function selects the appropriate values from the efo_labels table.
}
\examples{
getDiseaseLocation(c("EFO:0000094", "EFO:0000095"))
}
\author{
Robert Gentleman
}
