% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getMatchedTraits}
\alias{getMatchedTraits}
\title{A function to return the set of traits that were mapped to a given set of EFO terms}
\usage{
getMatchedTraits(Ontonames)
}
\arguments{
\item{Ontonames}{a character vector of the ontology CURIE symbols}
}
\value{
A named list, each element of which is named with the CURIE symbol, and the elements are the text
descriptions of the GWAS Catalog traits that mapped to that symbol.
}
\description{
This function provides an interface to the SQL table of matched traits to terms
}
\details{
The function extracts the matches, groups them by EFO CURIE label and returns a named list.
}
\examples{
getMatchedTraits(c("EFO:0000094", "EFO:0000095"))
}
\author{
Robert Gentleman
}
