% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getParents}
\alias{getParents}
\title{A function to query the efo_edges table and return the set of parent terms for the input ontology terms}
\usage{
getParents(Ontonames)
}
\arguments{
\item{Ontonames}{a character vector of the ontology CURIE symbols}
}
\value{
A named list, names are from the input Ontonames and the values are the set of parent terms.
}
\description{
This function provides an interface to the SQL database containing ontology edges. These are
}
\details{
The function selects the appropriate rows from the efo_edges table.
}
\examples{
getParents(c("EFO:0000094", "EFO:0000095"))
}
\author{
Robert Gentleman
}
