% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getSynonyms}
\alias{getSynonyms}
\title{A function to query the efo_synonyms table and return the set of known synonyms for the input ontology terms}
\usage{
getSynonyms(Ontonames)
}
\arguments{
\item{Ontonames}{a character vector of the ontology CURIE symbols}
}
\value{
A named list, each element of which is named with the CURIE symbol, and the elements are the text
of the synonyms
}
\description{
This function provides an interface to the SQL database containing ontology term synonyms. These are
synonyms provided as part of the ontology.
}
\details{
The function selects the appropriate rows and then splits them according to the input ontology terms.
Any ontology term without synonyms will be dropped.
}
\examples{
getSynonyms(c("EFO:0000094", "EFO:0000095"))
}
\author{
Robert Gentleman
}
