% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getUberonText}
\alias{getUberonText}
\title{A function to query the uberon_labels table and return the text descriptions for the input Uberon terms}
\usage{
getUberonText(UberonIDs)
}
\arguments{
\item{UberonIDs}{a character vector of the ontology CURIE symbols}
}
\value{
A named vector, names are from the input UberonIDs and the values are the corresponding text descriptions.
}
\description{
This function provides an interface to the SQL database containing Uberon labels.
}
\details{
The function selects the appropriate rows from the uberon_labels table.
}
\examples{
getUberonText(c("UBERON:0001004", "UBERON:0000474"))
}
\author{
Robert Gentleman
}
