% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getVersionInfo}
\alias{getVersionInfo}
\title{A function to return the version information from the Ontology/EBI databases}
\usage{
getVersionInfo()
}
\value{
A data.frame with one row per resource.
}
\description{
This function returns data.frame with one row for each resource that gives version information.
}
\details{
The version number for the resource, if it has one, is given. Otherwise the download
is given.
}
\examples{
# example code
getVersionInfo()
}
\author{
Robert Gentleman
}
