% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocgr_from_db.R
\name{granges_from_study}
\alias{granges_from_study}
\title{build a GRanges with all GWAS associations from SQLite db}
\usage{
granges_from_study(acc)
}
\arguments{
\item{acc}{character() study accession number(s) from EBI}
}
\value{
a data.frame with relevant information about hits reported in the selected studies
}
\description{
build a GRanges with all GWAS associations from SQLite db
}
\note{
Uses gwascatalog_associations table in the GWASCatalogSearchDB.
}
\examples{
tab = granges_from_study(acc="GCST007087")
}
