% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gwascatdb_path}
\alias{gwascatdb_path}
\title{get path to SQLite database for gwascatSearch, adding it (from Bioconductor Open Storage Network) to BiocFileCache if absent}
\usage{
gwascatdb_path(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache}
}
\value{
NULL if gwascatalog_search.db is already present in cache, otherwise named string
with path to SQLite database in cache
}
\description{
get path to SQLite database for gwascatSearch, adding it (from Bioconductor Open Storage Network) to BiocFileCache if absent
}
