% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gwasCatSearch_dbconn}
\alias{gwasCatSearch_dbconn}
\title{A function that allows the user to access the SQLite database tables provided with gwasCatSearch}
\usage{
gwasCatSearch_dbconn()
}
\description{
An instance of the SQLiteConnection class that has been opened on the database tables provided
}
\examples{
RSQLite::dbListTables(gwasCatSearch_dbconn())
}
