% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwc_df}
\alias{gwc_df}
\title{A data frame containing the GWAS catalog studies and EFO mappings}
\format{
A data frame with observations on
\describe{
\item{\code{STUDY.ACCESSION}}{The GWAS Catalog ID.}
\item{\code{DISEASE.TRAIT}}{The text description for the trait being studied.}
\item{\code{MAPPED_TRAIT}}{The text description of EFO term mapped to that trait.}
\item{\code{MAPPED_TRAIT_URI}}{The URI for the mapped EFO term.}
\item{\samp{MAPPED_TRAIT_CURIE}}{The CURIE for the mapped EFO term.}
}
}
\source{
EFO: https://www.ebi.ac.uk/efo/
}
\usage{
gwc_df
}
\description{
A data frame containing the GWAS catalog studies and EFO mappings
}
\note{
We use mappings from study defined traits to EFO as provided by the GWAS Catalog
}
\keyword{datasets}
