% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwc_gr}
\alias{gwc_gr}
\title{A restricted snapshot of the EBI GWAS catalog made 2023-08-28, as a GRanges instance}
\format{
GRanges instance with mcols columns
\describe{
\item{"PUBMEDID"}{a numeric vector}
\item{"MAPPED_TRAIT"}{a character vector}
\item{"MAPPED_GENE"}{a character vector}
\item{"STUDY ACCESSION"}{a character vector}
\item{"PVALUE_MLOG"}{a numeric vector}
}
}
\source{
This was created using makeCurrentGwascat in the Bioconductor gwascat package.
}
\usage{
gwc_gr
}
\description{
A restricted snapshot of the EBI GWAS catalog made 2023-08-28, as a GRanges instance
}
\examples{
data(gwc_gr)
tail(sort(table(S4Vectors::mcols(gwc_gr)$MAPPED_GENE)))
}
\keyword{datasets}
