% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwc_tc}
\alias{gwc_tc}
\title{A tCorpus object built using GWAS Catalog Studies.}
\format{
a tCorpus R6 object where documents are GWAS Catalog studies, suitable for searching
}
\source{
The data are created from \code{gwc_tc}
}
\usage{
gwc_tc
}
\description{
A tCorpus object built using GWAS Catalog Studies.
}
\examples{
gwc_tc
htg <- search_features(gwc_tc, query = "leuk*")
summary(htg)
}
\references{
GWAS catalog: https://www.ebi.ac.uk/gwas/
}
\seealso{
\link{efo_tc}, \code{\link[corpustools:tCorpus]{corpustools::tCorpus()}}
}
\keyword{datasets}
