% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GWCATID2PMID}
\alias{GWCATID2PMID}
\title{A function to query the gwascatalog_metadata table and return the set of PubMed IDs
that correspond to the input set of GWAS catalog Study Accession numbers.}
\usage{
GWCATID2PMID(GWCATID)
}
\arguments{
\item{GWCATID}{a character vector of the GWAS catalog Study Accession numbers.}
}
\description{
This function provides an interface to the SQL database containing the GWAS catalog metadata. Note that
one PubMed ID can give rise to many thousand GWAS catalog studies.
}
\details{
The function returns a vector PubMed IDs matching the input.
}
\examples{
gwid = GWCATID2PMID(c("GCST010478"))
length(gwid)
}
