% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits2DT.R
\name{hits2DTGWC}
\alias{hits2DTGWC}
\title{A function that takes the output of query against the GWAS Catalog corpus and creates a datatable object}
\usage{
hits2DTGWC(hits, gwc_df, tc)
}
\arguments{
\item{hits}{A \code{featureHits} object from the corpustools package.}

\item{gwc_df}{A dataframe containing the information about the GWAS catalog studies.}

\item{tc}{A \code{tCorpus} object that was built using the \code{gwc_df} dataframe.}
}
\value{
A \code{datatable} object with columns:
\describe{
\item{Study}{The GWAS study ID.}
\item{Text}{The set of CURIE IDs mapped to that trait.}
\item{DiseaseTrait}{The text label for the GWAS trait for the Study.}
\item{Others}{One column for each labeled search term in the call that created the \code{hits} object. If
those terms were labeled then the columns will have the same label. The value in the column is the actual string
that was matched on.}
\item{field}{If there is a column in the \code{tcorpus$tokens} table named \code{field}, the values in it
are aligned with the documents and returned. If there is no column named \code{field} then this column
is missing.}
}
}
\description{
Users query the \code{gwc_tc} corpus to find GWAS studies that match the query. Those hits can be processed
by \code{hits2DTGWC} to produce a datatable which then provides some amount of interactivity.
}
\details{
Once hits have been found, this function obtains that actual text string and field where
the match occurred.
}
\examples{
htl = search_features(gwc_tc, query="leuk*")
htlf = addField2Hits(htl, gwc_tc)

summary(htlf)
gwhitsDT <- hits2DTGWC(htlf, gwc_df, gwc_tc)
}
\seealso{
\code{\link[corpustools:tCorpus]{corpustools::tCorpus()}}, \link{efo_df}, \code{\link[DT:datatable]{DT::datatable()}}
}
\author{
R. Gentleman
}
