% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{make_oi}
\alias{make_oi}
\title{produce an ontology_index instance from GWASCatalogSearchDB sqlite
using only a vector of EFO curies}
\usage{
make_oi(tags)
}
\arguments{
\item{tags}{character() vector of \code{XXX:nnnnnnn} tags where
the XXX are prefixes of CURIE tags, such as \code{EFO} or \code{DOID}.}
}
\description{
produce an ontology_index instance from GWASCatalogSearchDB sqlite
using only a vector of EFO curies
}
\examples{
tags = c("DOID:10113", "DOID:7551", 
     "EFO:0000549", "EFO:0000584", "EFO:0000638", 
     "EFO:0000650", "EFO:0000694")
oo = make_oi(tags) 
oo
if (requireNamespace("ontoProc"))
   ontoProc::onto_plot2(oo, names(oo$name))
}
