% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oi}
\alias{oi}
\title{a static instance of an ontology_index representing the
EFO from the SQLite database}
\format{
ontologyIndex ontology_index instance
}
\usage{
oi
}
\description{
a static instance of an ontology_index representing the
EFO from the SQLite database
}
\note{
might be replaced by dynamic creation in .onLoad
}
\examples{
data("oi", package="gwasCatSearch")
oi
}
\keyword{datasets}
