% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTags.R
\name{plotTags}
\alias{plotTags}
\title{present a graphical view of selected ontology terms}
\usage{
plotTags(tags)
}
\arguments{
\item{tags}{character() vector of EFO tags}
}
\description{
present a graphical view of selected ontology terms
}
\examples{
tags = c("EFO:0003884", "EFO:0004531", "EFO:0005208", "EFO:0009909", "EFO:0021409", "EFO:0021433")
plotTags(tags)
}
