% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{process_annotated}
\alias{process_annotated}
\title{processtable}
\usage{
process_annotated(
  query = "(kidney AND disease)",
  include_subclasses = TRUE,
  direct_subclasses_only = FALSE,
  sels = NULL
)
}
\arguments{
\item{query}{character(1) corpustools::search_features query string}

\item{include_subclasses}{logical(1)}

\item{direct_subclasses_only}{logical(1)}

\item{sels}{numeric vector of indices to retain (mainly for row selections
in app)}
}
\description{
processtable
}
