% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{PubMed2GWASCAT}
\alias{PubMed2GWASCAT}
\title{A function to query the gwascatalog_metadata table and return the subset of it that corresponds
to the input PubMed IDs.}
\usage{
PubMed2GWASCAT(PMID)
}
\arguments{
\item{PMID}{a character vector of the PubMed IDs.}
}
\description{
This function provides an interface to the SQL database containing the GWAS catalog metadata.
}
\details{
The function returns the rows of the metadata table with PubMed IDs matching the input.
}
\examples{
pmdf = PubMed2GWASCAT(c("34662886"))
dim(pmdf)
}
