% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{query2table}
\alias{query2table}
\title{query2table}
\usage{
query2table(query = "(kidney AND disease)")
}
\arguments{
\item{query}{character(1) corpustools::search_features query string}
}
\description{
query2table
}
