% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources_annotated_with_term.R
\name{resources_annotated_with_term}
\alias{resources_annotated_with_term}
\title{A function that finds which phenotypes from a collection of phenotypes (The GWAS Catalog for this package)
map to specific terms in an ontology (EFO for this package)}
\usage{
resources_annotated_with_term(
  search_terms,
  include_subclasses = TRUE,
  direct_subclasses_only = FALSE
)
}
\arguments{
\item{search_terms}{character() A label for the term (EFO in this case) that is of interest.}

\item{include_subclasses}{logical(1) defaults to TRUE; if \code{TRUE} then all terms mapped to subclasses (children) of the term are included.}

\item{direct_subclasses_only}{logical(1) defaults to FALSE; If \code{TRUE} then only terms that map to direct children (first step only) are included.}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{STUDY.ACCESSION}{The GWAS catalog ID for the study.}
\item{PUBMEDID}{The PUBMEDID for the Study}
\item{DISEASE_TRAIT}{The GWAS Catalog text label for that trait.}
\item{MAPPED_TRAIT}{The text description/label of the ontology term.}
\item{MAPPED_TRAIT_CURIE}{The CURIE symbol for the mapped trait.}
\item{MAPPED_TRAIT_URI}{The URI for the ontology term.}
}
}
\description{
This function relies on a SQLite database that contains mappings from a set of phenotypes to one or more
ontologies.  The database contains information about the organization of the ontology and hence can find
phenotypes that map directly to a term or those that are inherited at a term by virtue of having mapped
directly to one of the chilren or other more specific descriptions.
}
\details{
The mapping of openGWAS traits/phenotypes to EFO ontology terms was performed by \code{text2term}. The results were
then organized, together with information on the sturcture of the ontology (EFO in this case) to facilitate retrieval.
Included in the return values is the confidence score provided by \code{text2term} which gives some indication of how syntactically
similar the phenotype description was to the ontology term label.
}
\examples{
ematch <- resources_annotated_with_term("EFO:0005297", include_subclasses = FALSE)
dim(ematch)
ematch[1, ]
em2 <- resources_annotated_with_term(c("EFO:1000830", "EFO:0004239"), include_subclasses=TRUE)
em3 <- resources_annotated_with_term(c("EFO:0004220", "EFO:1000049"), include_subclasses=FALSE)
}
\references{
\code{text2term}
}
\author{
Robert Gentleman
}
