% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{search_gwascat}
\alias{search_gwascat}
\title{Simple app for interrogating an ontology-mapped phenotype term corpus derived from EBI/NHGRI GWAS catalog}
\usage{
search_gwascat()
}
\value{
Used only for side effect of running a shiny app.
}
\description{
Simple app for interrogating an ontology-mapped phenotype term corpus derived from EBI/NHGRI GWAS catalog
}
\seealso{
A \href{https://vjcitn.shinyapps.io/gwasCatSearch}{shinyapps.io deployment} is available.
}
