% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ggmanh.R
\name{simple_ggmanh}
\alias{simple_ggmanh}
\title{produce plotly-capable Manhattan plot using ggplot2}
\usage{
simple_ggmanh(
  indf,
  pos.colname = "CHR_POS",
  chr.colname = "CHR_ID",
  pval.colname = "P-VALUE",
  label.colname = NULL,
  x.label = "Chromosome",
  y.label = "-log10 p",
  point.size = 0.75,
  signif = 5e-08
)
}
\arguments{
\item{indf}{data.frame as produced by \code{variants_from_study}}

\item{pos.colname}{character(1) defaults to "CHR_POS"}

\item{chr.colname}{character(1) defaults to "CHR_ID"}

\item{pval.colname}{character(1) defaults to "P-VALUE"}

\item{label.colname}{character(1) column used for hover tooltip labeling}

\item{x.label}{character(1) x-axis label}

\item{y.label}{character(1) y-axis label}

\item{point.size}{numeric(1) point size passed to geom_point}

\item{signif}{numeric(1) genome-wide significance threshold, default 5e-8}
}
\description{
produce plotly-capable Manhattan plot using ggplot2
}
\note{
Association records with missing p-values or positions are dropped before
plotting. The position column is coerced to numeric. No external Manhattan
plot package is required.
}
\examples{
con = gwasCatSearch_dbconn()
toprecs = DBI::dbGetQuery(con, "select * from gwascatalog_associations limit 200")
toprecs$CHR_POS = as.numeric(toprecs$CHR_POS)
ww = simple_ggmanh(toprecs, y.label="-log10 p",
       label.colname = "MAPPED_TRAIT", pval.colname="P-VALUE")
plotly::ggplotly(ww, tooltip="text")
}
