% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_from_study.R
\name{variants_from_study}
\alias{variants_from_study}
\title{given an accession, tabulate variants from GWAS catalog}
\usage{
variants_from_study(acc, gwcat)
}
\arguments{
\item{acc}{character() accession number(s) by EBI}

\item{gwcat}{instance of gwascat's gwaswloc class}
}
\value{
a data.frame with all relevant information about hits reported in the selected studies
}
\description{
given an accession, tabulate variants from GWAS catalog
}
\examples{
data(gwascat_2023_06_24)
tab = variants_from_study(acc="GCST007087", gwcat=gwascat_2023_06_24)
dim(tab)
head(tab[,1:8])
}
