% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwascat_2023_06_24}
\alias{gwascat_2023_06_24}
\title{A snapshot of EBI GWAS catalog}
\format{
an instance of gwaswloc from gwascat package
}
\source{
see gwascat package function makeCurrentGwascat
}
\usage{
gwascat_2023_06_24
}
\description{
A snapshot of EBI GWAS catalog
}
\keyword{datasets}
