% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldserver.R, R/server.R
\name{make_oi}
\alias{make_oi}
\title{produce an ontology_index instance from GWASCatalogSearchDB sqlite}
\usage{
make_oi(con)

make_oi(con)
}
\arguments{
\item{con}{SQLite connection via RSQLite/DBI dbConnect}
}
\description{
produce an ontology_index instance from GWASCatalogSearchDB sqlite

produce an ontology_index instance from GWASCatalogSearchDB sqlite
}
\note{
This is used in the server for the \code{search_gwascat()} app.

This is used in the server for the \code{search_gwascat()} app.
}
\examples{
rcon = gwasCatSearch:::.datacache$dbconn
efo = make_oi(rcon)
efo$name[1:5]
efo$children[1:3] # do not disconnect or check will error
rcon = gwasCatSearch:::.datacache$dbconn
efo = make_oi(rcon)
efo$name[1:5]
efo$children[1:3] # do not disconnect or check will error
}
