% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{process_annotated}
\alias{process_annotated}
\title{processtable}
\usage{
process_annotated(
  query = "(kidney AND disease)",
  include_subclasses = TRUE,
  direct_subclasses_only = FALSE,
  sels = NULL
)
}
\description{
processtable
}
