% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources_annotated_with_term.R
\name{resources_annotated_with_term}
\alias{resources_annotated_with_term}
\title{A function that finds which phenotypes from a collection of phenotypes (openGWAS for this package)
map to specific terms in an ontology (EFO for this package).}
\usage{
resources_annotated_with_term(
  search_terms,
  include_subclasses = TRUE,
  direct_subclasses_only = FALSE
)
}
\arguments{
\item{search_terms}{character() A label for the term (EFO in this case) that is of interest.}

\item{include_subclasses}{logical(1) defaults to TRUE; if \code{TRUE} then all terms mapped to subclasses (children) of the term are included.}

\item{direct_subclasses_only}{logical(1) defaults to FALSE; If \code{TRUE} then only terms that map to direct children (first step only) are included.}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{OpenGWAS ID}{The openGWAS ID for the trait.}
\item{OpenGWAS Trait}{The openGWAS text label for that trait.}
\item{Ontology Term}{The text description/label of the ontology term.}
\item{Ontology Term ID}{The label for the ontology term.}
\item{Mapping Confidence}{The confidence score from mapping the phenotype label to the ontology label.}
}
}
\description{
This function relies on a SQLite database that contains mappings from a set of phenotypes to one or more
ontologies.  The database contains information about the organization of the ontology and hence can find
phenotypes that map directly to a term or those that are inherited at a term by virtue of having mapped
directly to one of the chilren or other more specific descriptions.
}
\details{
The mapping of openGWAS traits/phenotypes to EFO ontology terms was performed by \code{text2term}. The results were
then organized, together with information on the sturcture of the ontology (EFO in this case) to facilitate retrieval.
Included in the return values is the confidence score provided by \code{text2term} which gives some indication of how syntactically
similar the phenotype description was to the ontology term label.
}
\examples{
ematch = resources_annotated_with_term("EFO:0009605", include_subclasses=FALSE)
dim(ematch)
ematch[1,]
}
\references{
\code{text2term}
}
\author{
Robert Gentleman
}
