% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{this is called by search_gwascat, also symlinked to inst/app2 for shinyapps usage}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{formal element for shiny server component}

\item{output}{formal element for shiny server component}

\item{session}{formal element for shiny server component}
}
\description{
this is called by search_gwascat, also symlinked to inst/app2 for shinyapps usage
}
\note{
Not exported.
}
