% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ggmanh.R
\name{simple_ggmanh}
\alias{simple_ggmanh}
\title{produce plotly-capable manhattanplot based on ggmanh}
\usage{
simple_ggmanh(
  x,
  chromosome = NULL,
  rescale = TRUE,
  rescale.ratio.threshold = 5,
  signif.rel.pos = 0.4,
  color.by.highlight = FALSE,
  label.colname = NULL,
  pval.colname = "P.VALUE",
  x.label = "Chromosome",
  y.label = "-log10 p",
  point.size = 0.75,
  label.font.size = 2,
  max.overlaps = 20,
  plot.title = ggplot2::waiver(),
  plot.subtitle = ggplot2::waiver(),
  plot.width = 10,
  plot.height = 5
)
}
\arguments{
\item{x}{MPdata instance from ggmanh::manhattan_data_preprocess}

\item{chromosome}{character(1) or NULL, should be removed for gwasCatSearch}

\item{rescale}{logical(1)}

\item{rescale.ratio.threshold}{numeric(1)}

\item{signif.rel.pos}{numeric(1)}

\item{color.by.highlight}{logical(1)}

\item{label.colname}{character(1) obligatory for phenotype labeling}

\item{pval.colname}{character(1)}

\item{x.label}{character(1)}

\item{y.label}{character(1)}

\item{point.size}{numeric(1)}

\item{label.font.size}{numeric(1)}

\item{max.overlaps}{numeric(1)}

\item{plot.title}{like ggplot2::waiver() or character(1)}

\item{plot.subtitle}{like plot.title}

\item{plot.width}{numeric(1)}

\item{plot.height}{numeric(1)}
}
\description{
produce plotly-capable manhattanplot based on ggmanh
}
\note{
Association records with missing p-values are eliminated before
plotting.
}
\examples{
data(fdat)
dd = ggmanh::manhattan_data_preprocess(fdat, chr.colname="seqnames",
   pos.colname="start", pval.colname="P.VALUE", chr.order=c(1:22, "X", "Y"))
ww = simple_ggmanh(dd, y.label="-log10 p", 
       label.colname = "MAPPED_TRAIT", pval.colname="P.VALUE")
plotly::ggplotly(ww, tooltip="text")
}
