% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_from_study.R
\name{variants_from_study}
\alias{variants_from_study}
\title{given a set of Study Accession IDs extract the corresponding variants from GWAS catalog}
\usage{
variants_from_study(acc)
}
\arguments{
\item{acc}{character() accession number(s) by EBI}
}
\value{
a data.frame with relevant information about hits reported in the selected studies
}
\description{
given a set of Study Accession IDs extract the corresponding variants from GWAS catalog
}
\examples{
tab = variants_from_study(acc="GCST007087")
dim(tab)
head(tab)
}
