% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_from_study.R
\name{variants_from_study}
\alias{variants_from_study}
\title{given an accession, tabulate variants from GWAS catalog}
\usage{
variants_from_study(acc, gwdat)
}
\arguments{
\item{acc}{character() accession number(s) by EBI}

\item{gwdat}{instance of GRanges with \code{STUDY.ACCESSION} and other locus-level metadata}
}
\value{
a data.frame with relevant information about hits reported in the selected studies
}
\description{
given an accession, tabulate variants from GWAS catalog
}
\examples{
gwc_gr = gwasCatSearch:::.datacache$gwc_gr
tab = variants_from_study(acc="GCST007087", gwdat=gwc_gr)
dim(tab)
head(tab)
}
