% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_variant_context.R
\name{view_variant_context}
\alias{view_variant_context}
\title{interactive visualization of GWAS hits near a position}
\usage{
view_variant_context(
  chr = 15,
  pos = 6.9e+07,
  radius = 5e+05,
  focal_rec = NULL,
  gwdat,
  main = "Manhattan plot for GWAS catalog hits, mouseover for details"
)
}
\arguments{
\item{chr}{chromosome identifier}

\item{pos}{numeric(1) position of variant of interest}

\item{radius}{numeric(1) radius of interval to present}

\item{focal_rec}{NULL or data.frame with one row with elements CHR_ID, CHR_POS, PVALUE_MLOG, STUDY.ACCESSION,
MAPPED_TRAIT, SNP_ID_CURRENT; this will be used to identify and annotate the "focal"
SNP for the visualization with a blue dot}

\item{gwdat}{instance of GRanges with GWAS catalog p-values etc.}

\item{main}{character(1) title for plot}
}
\value{
ggplotly is called
}
\description{
interactive visualization of GWAS hits near a position
}
\examples{
gwc_gr = gwasCatSearch:::.datacache$gwc_gr
foc = S4Vectors::mcols(IRanges::subsetByOverlaps(gwc_gr,
     GenomicRanges::GRanges("15:69287238")))
foc = as.data.frame(foc)  # will fix names
foc$CHR_POS = 69287238
foc$CHR_ID = 15
view_variant_context(gwdat=gwc_gr, focal_rec=foc)
}
