% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_variant_context.R
\name{view_variant_context}
\alias{view_variant_context}
\title{interactive visualization of GWAS hits near a position}
\usage{
view_variant_context(
  chr = 15,
  pos = 6.9e+07,
  radius = 5e+05,
  gwcat,
  main = "Manhattan plot for GWAS catalog hits, mouseover for details"
)
}
\arguments{
\item{chr}{chromosome identifier}

\item{pos}{numeric(1) position of variant of interest}

\item{radius}{numeric(1) radius of interval to present}

\item{gwcat}{instance of gwaswloc from gwascat package}

\item{main}{character(1) title for plot}
}
\value{
ggplotly is called
}
\description{
interactive visualization of GWAS hits near a position
}
\examples{
data(gwascat_2023_06_24)
view_variant_context(gwcat=gwascat_2023_06_24)
}
